requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
azure-core==1.32.0
azure-storage-blob==12.24.1
backports.zoneinfo==0.2.1
billiard==4.2.1
boto3==1.36.13
botocore==1.36.13
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Django==4.2.19
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
isodate==0.7.2
jmespath==1.0.1
kombu==5.4.2
mssql-django==1.5
numpy==1.24.4
pandas==2.0.3
pillow==10.4.0
pkg_resources==0.0.0
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
redis==5.2.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
